backend/requirements.txt
fastapi
uvicorn
celery
redis
psycopg2-binary
sqlalchemy
websockets
pyjwt
pandas
scikit-learn
joblib
pydantic
httpx
python-dotenv
razorpay
passlib[bcrypt]
